tests/wpt/mozilla/tests/mozilla/selectionchange/selectionchange-selection.html

<html>
<head>
  <title>Test that selectionchange fires when Selection methods cause changes</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
</head>
<body>
<span id="abcd">abcd</span><span id="efgh">efgh</span><span id="ijkl">ijkl</span>
<script>
var t = async_test("Selection methods fire selectionchange");
var r = new Range();
var abcd = document.getElementById("abcd");
var efgh = document.getElementById("efgh");
var ijkl = document.getElementById("ijkl");
r.setStart(abcd, 0);
r.setEnd(ijkl, 1);
var s = document.getSelection();
s.addRange(r);
// A task is now queued that will fire a selectionchange event,
// so the following listener will be called soon, even though it wasn't
// listening at the time we added the range.
var step = 0;
document.onselectionchange = () => {
  t.step(() => {
    switch(step++) {
      // order chosen so that s.type always changes between consecutive steps
      case 0: assert_equals(s.type, "Range"); s.removeRange(r); break;
      case 1: assert_equals(s.type, "None"); s.collapse(efgh, 0); break;
      case 2: assert_equals(s.type, "Caret"); s.removeAllRanges(); break;
      case 3: assert_equals(s.type, "None"); s.setPosition(efgh, 1); break;
      case 4: assert_equals(s.type, "Caret"); s.extend(ijkl, 1); break;
      case 5: assert_equals(s.type, "Range"); s.collapseToStart(); break;
      case 6: assert_equals(s.type, "Caret"); s.setBaseAndExtent(abcd, 0, efgh, 0); break;
      case 7: assert_equals(s.type, "Range"); s.collapseToEnd(); break;
      case 8: assert_equals(s.type, "Caret"); s.empty(); break;
      case 9: assert_equals(s.type, "None"); s.selectAllChildren(efgh); break;
      case 10: assert_equals(s.type, "Range"); s.deleteFromDocument(); break;
      case 11: assert_equals(s.type, "Caret"); queueFinish(); break;
      case 12: assert_unreached("Too many selectionchange events");
    }
  },"Step number "+step);
}
function queueFinish() {
  // Finish slightly later than the last selectionchange task,
  // so if there are any extra ones queued we have time to
  // hit the assert_unreached.
  setTimeout(() => { t.done(); }, 1);
}
</script>
</body>
</html>
